:recycle: refactor: Refactor DrawManager.java
