+++ src/screen/HighScoreScreen.java
@@ -13,6 +13,8 @@ import engine.SoundManager;
  *
  * @author <a href="mailto:RobertoIA1987@gmail.com">Roberto Izquierdo Amo</a>
  */
+@SuppressWarnings({"PMD.LawOfDemeter"})
+
 public class HighScoreScreen extends Screen {
 
   /** List of past high scores. */
@@ -77,7 +79,7 @@ public class HighScoreScreen extends Screen {
     if (inputManager.isMouseClicked()) {
       final int mx = inputManager.getMouseX();
       final int my = inputManager.getMouseY();
-      final java.awt.Rectangle backBox = drawManager.getBackButtonHitbox(this);
+      final java.awt.Rectangle backBox = drawManager.menu().getBackButtonHitbox(this);
 
       if (backBox.contains(mx, my)) { // NOPMD - LawOfDemeter
         this.returnCode = 1;
@@ -94,9 +96,9 @@ public class HighScoreScreen extends Screen {
   private void draw() {
     drawManager.initDrawing(this);
 
-    drawManager.drawHighScoreMenu(this);
-    drawManager.drawHighScores(this, getPlayerScores("1P"), "1P"); // Left column
-    drawManager.drawHighScores(this, getPlayerScores("2P"), "2P"); // Right column
+    drawManager.menu().drawHighScoreMenu(this);
+    drawManager.menu().drawHighScores(this, getPlayerScores("1P"), "1P"); // Left column
+    drawManager.menu().drawHighScores(this, getPlayerScores("2P"), "2P"); // Right column
 
     // hover highlight
     handleBackButtonHover();
